+++ src/lib/services/amplitude.tsx
@@ -1,7 +1,7 @@
 import { track } from "@amplitude/analytics-browser";
 
 import type { AttachFundsType } from "lib/components/fund/types";
-import type { Option } from "lib/types";
+import type { Dict, Option } from "lib/types";
 
 export enum AmpEvent {
   INVALID_STATE = "To Invalid State",
@@ -35,7 +35,7 @@ export enum AmpEvent {
   TO_DEPLOY = "To Deploy",
   TO_UPLOAD = "To Upload",
   TO_INSTANTIATE = "To Instantiate",
-  TO_PROPOSALS = "To Proposals",
+  TO_PROPOSAL_LIST = "To Proposal List",
   TO_QUERY = "To Query",
   TO_EXECUTE = "To Execute",
   TO_MIGRATE = "To Migrate",
@@ -96,6 +96,13 @@ export enum AmpEvent {
   USE_UNSUPPORTED_ASSETS = "Use Unsupported Assets",
   USE_TX_MSG_EXPAND = "Use Transaction Message Expand",
   USE_EXPAND = "Use General Expand",
+  USE_SEARCH_INPUT = "Use Search Input",
+  USE_FILTER_MY_PROPOSALS = "Use Filter My Proposals",
+  USE_FILTER_PROPOSALS_STATUS = "Use Filter Proposals Status",
+  USE_FILTER_PROPOSALS_TYPE = "Use Filter Proposals Types",
+  USE_PAGINATION_PAGE_SIZE = "Use Pagination Page Size",
+  USE_PAGINATION_NAVIGATION = "Use Pagination Navigation",
+  USE_CREATE_NEW_PROPOSAL = "Use Create New Proposal",
   // TX
   TX_SUCCEED = "Tx Succeed",
   TX_FAILED = "Tx Failed",
@@ -131,7 +138,10 @@ type SpecialAmpEvent =
   | AmpEvent.USE_VIEW_JSON
   | AmpEvent.USE_UNSUPPORTED_ASSETS
   | AmpEvent.USE_COPIER
-  | AmpEvent.USE_EXPAND;
+  | AmpEvent.USE_EXPAND
+  | AmpEvent.USE_PAGINATION_NAVIGATION
+  | AmpEvent.USE_FILTER_PROPOSALS_STATUS
+  | AmpEvent.USE_FILTER_PROPOSALS_TYPE;
 
 export const AmpTrackInvalidState = (title: string) =>
   track(AmpEvent.INVALID_STATE, { title });
@@ -174,8 +184,11 @@ export const AmpTrackUseRadio = (radio: string) =>
 export const AmpTrackUseOtherModal = (title: string) =>
   track(AmpEvent.USE_OTHER_MODAL, { title });
 
-export const AmpTrackMintscan = (type: string) =>
-  track(AmpEvent.MINTSCAN, { type });
+export const AmpTrackMintscan = (
+  type: string,
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
+  properties?: Dict<string, any>
+) => track(AmpEvent.MINTSCAN, { type, properties });
 
 export const AmpTrackWebsite = (url: string) =>
   track(AmpEvent.WEBSITE, { url });
@@ -199,3 +212,20 @@ export const AmpTrackExpand = (
   component: "assets" | "json" | "permission_address" | "event_box",
   section: Option<string>
 ) => track(AmpEvent.USE_EXPAND, { action, component, section });
+
+export const AmpTrackUseFilter = (
+  ampEvent: AmpEvent,
+  filters: string[],
+  action: string
+) => track(ampEvent, { action, filters });
+
+export const AmpTrackPaginationNavigate = (
+  navigate: string,
+  pageSize: number,
+  currentPage: number
+) =>
+  track(AmpEvent.USE_PAGINATION_NAVIGATION, {
+    navigate,
+    pageSize,
+    currentPage,
+  });